+++ src/commands/default.ts
@@ -1,10 +1,11 @@
 import { readFile, writeFile } from 'fs';
 import { Templater } from "@guscrawford.com/json-xform";
-import { read, parseVars } from '../util';
-import { RegisteredCommand, CliApp, StaticOption, Command, StaticArgument } from "@guscrawford.com/cleye";
+import { read, parseVars, stripComments } from '../util';
+import { RegisteredCommand, CliApp, StaticOption, Command, StaticArgument, StaticCommand, Option } from "@guscrawford.com/cleye";
 import { join } from "path";
-import { CliAppInstance } from "../xform-cli";
-export class DefaultCommand implements Command {
+import { default as fromJson } from 'jsontoxml';
+const jsonYml = require('json2yaml');
+export class DefaultCommand implements StaticCommand {
     name = "default";
     options: {[key:string]:StaticOption} = {
         'out':<StaticOption>{
@@ -18,6 +19,18 @@ export class DefaultCommand implements Command {
         'extends':<StaticOption>{
             spinalCaseName:'extends',
             flag:'E'
+        },
+        'remove':<StaticOption>{
+            spinalCaseName:'remove',
+            flag:'R'
+        },
+        'format':<StaticOption>{
+            spinalCaseName:'format',
+            flag:'F'
+        },
+        'comments':<StaticOption>{
+            spinalCaseName:'comments',
+            flag:'C'
         }
     };
     args: {[key:string]:StaticArgument} = {
@@ -25,34 +38,43 @@ export class DefaultCommand implements Command {
             name:'template'
         }
     };
-    index:number = -1;
-    registeredCommand:RegisteredCommand =  this.app.command(this);
-    constructor(protected app:CliApp) {  }
+    run = main_default;
 }
-export const RegisteredDefaultCommand = new DefaultCommand(CliAppInstance).registeredCommand;
-let fileArg:string, relPath:string[], workingDirectory:string;
-if ((CliAppInstance as any).cancelDefault !== true) {
-    fileArg = join(process.cwd(),(RegisteredDefaultCommand.args.template as any).value);
+function main_default(app:CliApp) {
+    let fileArg:string, relPath:string[], workingDirectory:string;
+    fileArg = join(process.cwd(),((app.commands.default as Command).args.template as any).value);
     relPath = fileArg.split(/\/|\\/);
     relPath.pop();
     if (!fileArg.match(/^\/|([a-zA-Z]\:)?\\/)) 
         workingDirectory = join(process.cwd(),relPath.join('/'));
     else workingDirectory = relPath.join('/');
-    if ((RegisteredDefaultCommand.args.template as any).value)
-        readFile(fileArg,{encoding:'utf8'},transform(RegisteredDefaultCommand));
-    else if (!(RegisteredDefaultCommand.args.template as any).value)
-        read(process.stdin, transform(RegisteredDefaultCommand));
+    if (((app.commands.default as Command).args.template as any).value)
+        readFile(
+            join(process.cwd(),((app.commands.default as Command).args.template as any).value),
+            {encoding:'utf8'},
+            transform(app.commands.default as RegisteredCommand, workingDirectory)
+        );
+    else if (!((app.commands.default as Command).args.template as any).value)
+        read(process.stdin, transform(app.commands.default as RegisteredCommand, workingDirectory));
 }
-const templateWithVars = (data:string) => parseVars(
+const templateWithVars = (data:string, varOption:Option) => parseVars(
     JSON.parse(data),
-    (RegisteredDefaultCommand.options.var as any).value
+    varOption.value as string
 );
-export function transform(command:RegisteredCommand) { return (err:any, data:string) => {
+function transform(command:RegisteredCommand, workingDirectory:string) { return (err:any, data:string) => {
     if (err) throw err;
+    let allowComments = !command.options.comments || command.options.comments && (command.options.comments as any).value !== 'false';
+    if (allowComments) data = stripComments(data);
     try {
-        let parsedData = parse(command)(data);
+        let parsedData = parse(command, workingDirectory)(data);
+        let yaml = command.options.format.value && (command.options.format as any).value.match(/ya?ml/i) || ((command.options.out as any).value && ((command.options.out as any).value as string).endsWith('.yml'));
+        let xml = command.options.format.value && (command.options.format as any).value.match(/xml/i) || ((command.options.out as any).value && ((command.options.out as any).value as string).endsWith('.xml'));
+        
+        if (xml) parsedData = fromJson(parsedData, {removeIllegalNameCharacters:true});
+        else if (yaml) parsedData = jsonYml.stringify(parsedData);
+        else parsedData = asString(parsedData);
         if (!(command.options.out as any).value)
-            process.stdout.write(Buffer.from(parsedData+"\n\r"));
+            process.stdout.write(Buffer.from(parsedData)+"\n");
         else
             writeFile(
                 join(process.cwd(),(command.options.out as any).value||(command.args.template as any).value), 
@@ -63,20 +85,20 @@ export function transform(command:RegisteredCommand) { return (err:any, data:str
         throw e;
     }
 }; }
-export function parse(command:RegisteredCommand) {
+function parse(command:RegisteredCommand, workingDirectory:string) {
     return (data:string)=> {
-        // console.error(new Templater({"@xform:extends":'example/sample.json'}).parse())
         var template = (command.options.extends as any).value
-           ? Object.assign(templateWithVars(data), {"@xform:extends":(command.options.extends as any).value})
-           : templateWithVars(data);
-        // console.error(template);
+           ? Object.assign(templateWithVars(data, command.options.var as Option), {"@xform:extends":(command.options.extends as any).value})
+           : templateWithVars(data, command.options.var as Option);
         var templater = new Templater(template,undefined,workingDirectory);
         var parsedTemplate = templater.parse();
-        // console.error(templater);
-        return JSON.stringify(
-            parsedTemplate,
-            null,
-            "  "
-        );
+        return parsedTemplate;
     }
+}
+function asString(parsedTemplate:any) {
+    return JSON.stringify(
+        parsedTemplate,
+        null,
+        "  "
+    );
 }